Datatype Change.

+++ database/migrations/2018_02_21_204413_create_sub_questions_table.php
@@ -16,7 +16,7 @@ class CreateSubQuestionsTable extends Migration
         Schema::create('sub_questions', function (Blueprint $table) {
             $table->increments('id');
             $table->text('question');
-            $table->integer('max_marks');
+            $table->float('max_marks');
             $table->integer('exam_question_id');
             $table->text('co');
             $table->timestamps();
